all node integration test in single file
